+++ services/web/server/src/simcore_service_webserver/projects/_read_utils.py
@@ -12,7 +12,7 @@ from servicelib.utils import logged_gather
 from simcore_postgres_database.webserver_models import ProjectType as ProjectTypeDB
 from simcore_service_webserver.rest_schemas_base import OutputSchema
 
-from ..catalog import catalog_plugin
+from ..catalog.client import get_services_for_user_in_product
 from . import projects_api
 from ._permalink import update_or_pop_permalink_in_project
 from ._rest_schemas import ProjectListItem
@@ -56,9 +56,7 @@ async def list_projects(
     app = request.app
     db = ProjectDBAPI.get_from_app_context(app)
 
-    user_available_services: list[
-        dict
-    ] = await catalog_plugin.get_services_for_user_in_product(
+    user_available_services: list[dict] = await get_services_for_user_in_product(
         app, user_id, product_name, only_key_versions=True
     )
