+++ services/web/server/tests/unit/isolated/test_catalog_setup.py
@@ -10,11 +10,11 @@ from aiohttp.test_utils import TestClient
 from servicelib.aiohttp.application import create_safe_application
 from servicelib.aiohttp.client_session import APP_CLIENT_SESSION_KEY
 from simcore_service_webserver._meta import api_version_prefix
-from simcore_service_webserver.catalog.catalog_plugin import (
+from simcore_service_webserver.catalog.client import to_backend_service
+from simcore_service_webserver.catalog.plugin import (
     is_catalog_service_responsive,
     setup_catalog,
 )
-from simcore_service_webserver.catalog.client import to_backend_service
 from simcore_service_webserver.rest import APP_OPENAPI_SPECS_KEY, load_openapi_specs
 from yarl import URL
