+++ services/web/server/tests/unit/with_dbs/01/studies_dispatcher/test_studies_dispatcher_handlers.py
@@ -23,7 +23,7 @@ from pytest_simcore.helpers.utils_login import UserRole
 from pytest_simcore.pydantic_models import iter_model_examples_in_module
 from servicelib.json_serialization import json_dumps
 from settings_library.redis import RedisSettings
-from simcore_service_webserver.catalog import catalog_plugin
+from simcore_service_webserver.catalog import plugin as catalog_plugin
 from simcore_service_webserver.studies_dispatcher._core import ViewerInfo
 from simcore_service_webserver.studies_dispatcher._rest_handlers import ServiceGet
 from sqlalchemy.sql import text
